requirements.txt
numpy
matplotlib
scikit-learn
finch-genetics==3.5.3